feat: 06장 pod irsa 작업중

+++ 06_Workload_Access_2/01_app_irsa/03_irsa_create_policy.sh
@@ -13,6 +13,13 @@ if [ ! -d "tmp" ]; then
     mkdir -p tmp
 fi
 
+# Check if policy already exists
+POLICY_EXISTS=$(aws iam list-policies --query "Policies[?PolicyName=='${POLICY_NAME}'].Arn" --output text ${PROFILE_STRING})
+
+if [ ! -z "$POLICY_EXISTS" ]; then
+    echo "${POLICY_NAME} policy 가 존재합니다."
+    exit 0
+fi
 cat >tmp/irsa-workload-policy.json <<EOF
 {
     "Version": "2012-10-17",
@@ -29,5 +36,10 @@ EOF
 echo "aws iam create-policy --policy-name ${POLICY_NAME} \\
     --policy-document file://tmp/irsa-workload-policy.json ${PROFILE_STRING}"
 
+echo "${POLICY_NAME} 생성중"
 aws iam create-policy --policy-name ${POLICY_NAME} \
     --policy-document file://tmp/irsa-workload-policy.json ${PROFILE_STRING}
+
+aws iam wait policy-exists \
+    --policy-arn arn:aws:iam::${AWS_ACCOUNT_ID}:policy/${POLICY_NAME} ${PROFILE_STRING}
+echo "${POLICY_NAME} 생성 완료"
